fix: allow login by email case-insensitive (#7578)
A customer noted that the login by email was case-sensitive, which differs to the handling of the loginname.

This PR changes the email check to be case-insensitive (which it was already in same parts) and improve the search for this as well.

(cherry picked from commit 51763d105d4eb4a30965f4539f3e6353cca3e7b6)