docs: copy and inline logging config (#8415)
# Which Problems Are Solved

Links to the logging config in the github repo point to the wrong line
numbers. When they are fixed, it is highly likely that they get outdated
again very fast.

# How the Problems Are Solved

The logging relevant config section is copied to the docs page.
This copy is less likely to get outdated very fast, because the
`LogStore` sections content is not expected to change a lot but its
position is.

# Additional Changes

Adds the header *Why ZITADEL does not write logs to files*

# Additional Context

Reported by a customer in a support request.