docs(cloud login v2): use default login url (#9477)
# Which Problems Are Solved

We configured the default base URL for the hosted v2 login to
`/ui/v2/login`. However, the docs still instruct readers to configure
the URL explicitly. This is unneccesary mental overhead and a risk of
self-DOS due to typos.

# How the Problems Are Solved

The docs instruct readers to not configure the base URL in order to use
the default.

Co-authored-by: Livio Spring <livio.a@gmail.com>